Using Optimizations details from Koku-ui microfrontend
